fixed sorting bug in my submissions page
