+++ frontend/pages/submissions.js
@@ -1,17 +1,11 @@
-import {
-  faGlobeAmericas,
-  faShareAlt,
-  faUserLock
-} from '@fortawesome/free-solid-svg-icons';
-import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 import axios from 'axios';
-import { useRouter } from 'next/router';
 import { useEffect, useState } from 'react';
 import { useSelector } from 'react-redux';
 import useSWR from 'swr';
 
 import Basket from '../components/Basket';
 import Table from '../components/ReusableComponents/Table/Table';
+import SubmissionDisplay from '../components/SubmissionComponents/SubmissionDisplay';
 import TableButtons from '../components/SubmissionComponents/TableButtons';
 import TopLevel from '../components/TopLevel';
 import styles from '../styles/submissions.module.css';
@@ -115,91 +109,6 @@ function Submissions() {
   );
 }
 
-function SubmissionDisplay(properties) {
-  const router = useRouter();
-
-  const [privacyDisplay, setPrivacyDisplay] = useState();
-
-  useEffect(() => {
-    if (properties.submission.privacy === 'public')
-      setPrivacyDisplay(<FontAwesomeIcon icon={faGlobeAmericas} size="1x" />);
-    else if (properties.submission.privacy === 'private')
-      setPrivacyDisplay(
-        <FontAwesomeIcon icon={faUserLock} color="#ff0000" size="1x" />
-      );
-    else setPrivacyDisplay(<FontAwesomeIcon icon={faShareAlt} size="1x" />);
-  }, [properties.submission.privacy]);
-
-  return (
-    <tr
-      key={properties.submission.displayId}
-      className={styles.submission}
-      onClick={() => {
-        router.push(
-          properties.submission.uri.replace(
-            'https://synbiohub.org',
-            process.env.backendUrl
-          )
-        );
-      }}
-    >
-      <td>
-        <input
-          checked={properties.selected.get(properties.submission.displayId)}
-          onChange={event => {
-            properties.setSelected(
-              new Map(
-                properties.selected.set(
-                  properties.submission.displayId,
-                  event.target.checked
-                )
-              )
-            );
-          }}
-          onClick={event => {
-            event.stopPropagation();
-          }}
-          type="checkbox"
-        />
-      </td>
-      <td>
-        <code>{properties.submission.name}</code>
-      </td>
-      <td>{properties.submission.displayId}</td>
-      <td>{properties.submission.description}</td>
-      <td>{properties.submission.type}</td>
-      <td>{privacyDisplay}</td>
-    </tr>
-  );
-}
-
-const processSubmissions = submissions => {
-  for (const submission of submissions) {
-    const potentialType = submission.type.toLowerCase();
-
-    // Identify what type of object the search result is from type url
-    if (potentialType.includes('component')) {
-      submission.type = 'Component';
-    }
-    if (potentialType.includes('sequence')) {
-      submission.type = 'Sequence';
-    }
-    if (potentialType.includes('module')) {
-      submission.type = 'Module';
-    }
-    if (potentialType.includes('collection')) {
-      submission.type = 'Collection';
-    }
-
-    if (submission.privacy !== 'shared') {
-      submission.privacy = 'public';
-      if (!submission.uri.includes('/public/')) submission.privacy = 'private';
-    }
-  }
-
-  return submissions;
-};
-
 const options = [
   { value: 'name', label: 'Name' },
   { value: 'displayId', label: 'Display ID' },
@@ -207,15 +116,19 @@ const options = [
   { value: 'privacy', label: 'Privacy' }
 ];
 
-const sortString = (plugin1, plugin2) => {
-  return (plugin1.name > plugin2.name && 1) || -1;
+const compareStrings = (string1, string2) => {
+  return (string1.toLowerCase() > string2.toLowerCase() && 1) || -1;
 };
 
 const sortMethods = {
-  name: sortString,
-  displayId: sortString,
-  type: sortString,
-  privacy: sortString
+  name: (submission1, submission2) =>
+    compareStrings(submission1.name, submission2.name),
+  displayId: (submission1, submission2) =>
+    compareStrings(submission1.displayId, submission2.displayId),
+  type: (submission1, submission2) =>
+    compareStrings(submission1.type, submission2.type),
+  privacy: (submission1, submission2) =>
+    compareStrings(submission1.privacy, submission2.privacy)
 };
 
 const useMySubmissions = token => {
@@ -242,6 +155,33 @@ const useSharedSubmissions = token => {
     isSharedLoading: !error && !data,
     isSharedError: error
   };
+};
+
+const processSubmissions = submissions => {
+  for (const submission of submissions) {
+    const potentialType = submission.type.toLowerCase();
+
+    // Identify what type of object the search result is from type url
+    if (potentialType.includes('component')) {
+      submission.type = 'Component';
+    }
+    if (potentialType.includes('sequence')) {
+      submission.type = 'Sequence';
+    }
+    if (potentialType.includes('module')) {
+      submission.type = 'Module';
+    }
+    if (potentialType.includes('collection')) {
+      submission.type = 'Collection';
+    }
+
+    if (submission.privacy !== 'shared') {
+      submission.privacy = 'public';
+      if (!submission.uri.includes('/public/')) submission.privacy = 'private';
+    }
+  }
+
+  return submissions;
 };
 
 export default function SubmissionsWrapped() {
